Clean up unused files and organize navigation routes
- Move Gallery and Detail navigation classes to separate Routes.kt file
- Delete unused NavigationUtils files (all platform implementations)
- Remove unused Greeting.kt template file
- Delete GetCatInfoUseCase as it was never used
- Remove CatInfo data class and related functionality
- Update imports and DI module to reflect removed classes
- Simplify CatViewModel by removing unused use case dependency

This cleanup reduces the codebase size and removes confusing unused code.
