--- composeApp/src/commonMain/kotlin/org/godslew/gkcatinfosampleapp/Greeting.kt
@@ -1,9 +1,0 @@
-package org.godslew.gkcatinfosampleapp
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
