--- composeApp/src/commonMain/kotlin/org/godslew/gkcatinfosampleapp/navigation/NavigationUtils.kt
@@ -1,4 +1,0 @@
-package org.godslew.gkcatinfosampleapp.navigation
-
-expect fun encodeUrl(value: String): String
-expect fun decodeUrl(value: String): String
