+++ composeApp/src/commonMain/kotlin/org/godslew/gkcatinfosampleapp/navigation/Routes.kt
@@ -1,0 +1,13 @@
+package org.godslew.gkcatinfosampleapp.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object Gallery
+
+@Serializable
+data class Detail(
+    val imageId: String,
+    val imageUrl: String,
+    val breedId: String = ""
+)
